INS-19: tests split
