+++ functest/register_node_test.go
@@ -32,18 +32,9 @@ var scheme = platformpolicy.NewPlatformCryptographyScheme()
 var keyProcessor = platformpolicy.NewKeyProcessor()
 
 const TESTPUBLICKEY = "some_fancy_public_key"
-const NOTEXISTINGPUBLICKEY = "not_existing_public_key"
 
 func registerNodeSignedCall(params ...interface{}) (string, error) {
 	res, err := signedRequest(&root, "RegisterNode", params...)
-	if err != nil {
-		return "", err
-	}
-	return res.(string), nil
-}
-
-func getNodeRefSignedCall(params ...interface{}) (string, error) {
-	res, err := signedRequest(&root, "GetNodeRef", params...)
 	if err != nil {
 		return "", err
 	}
@@ -125,37 +116,4 @@ func TestReceiveNodeCert(t *testing.T) {
 			require.True(t, verified)
 		})
 	}
-}
-
-func TestGetNodeRefByPK(t *testing.T) {
-	const testRole = "light_material"
-	ref, err := registerNodeSignedCall(TESTPUBLICKEY, testRole)
-	require.NoError(t, err)
-	require.NotNil(t, ref)
-
-	nodeRef, err := getNodeRefSignedCall(TESTPUBLICKEY)
-	require.NoError(t, err)
-	require.Equal(t, ref, nodeRef)
-}
-
-func TestGetNodeRefByNotExistsPK(t *testing.T) {
-	const testRole = "light_material"
-	ref, err := registerNodeSignedCall(TESTPUBLICKEY, testRole)
-	require.NoError(t, err)
-	require.NotNil(t, ref)
-
-	nodeRef, err := getNodeRefSignedCall(NOTEXISTINGPUBLICKEY)
-	require.Equal(t, "", nodeRef)
-	require.Contains(t, err.Error(), "[ GetNodeRefByPK ] Node not found by PK: ")
-}
-
-func TestGetNodeRefInvalidParams(t *testing.T) {
-	const testRole = "light_material"
-	ref, err := registerNodeSignedCall(TESTPUBLICKEY, testRole)
-	require.NoError(t, err)
-	require.NotNil(t, ref)
-
-	nodeRef, err := getNodeRefSignedCall(123)
-	require.Equal(t, "", nodeRef)
-	require.Contains(t, err.Error(), "[ getNodeRef ] Can't unmarshal params: ")
 }
